Frontend updates

+++ energy-trading-rust/src/database/models.rs
@@ -1,18 +1,19 @@
 use serde::{Deserialize, Serialize};
 use chrono::{DateTime, Utc};
 use sqlx::FromRow;
+use bigdecimal::BigDecimal;
 
 #[derive(Debug, Clone, Serialize, Deserialize, FromRow)]
 pub struct Battery {
     pub id: i32,
     pub device_id: i32,
     pub owner_pubkey: String,
-    pub energy_total: f64,
-    pub percentage_for_sale: f64,
-    pub reserve_price: f64,
+    pub energy_total: BigDecimal,
+    pub percentage_for_sale: BigDecimal,
+    pub reserve_price: BigDecimal,
     pub health_status: i32,
-    pub voltage: f64,
-    pub discharge_rate: f64,
+    pub voltage: BigDecimal,
+    pub discharge_rate: BigDecimal,
     pub status: String,
     pub last_seen: DateTime<Utc>,
     pub created_at: DateTime<Utc>,
@@ -23,8 +24,8 @@ pub struct Aggregator {
     pub id: i32,
     pub device_id: i32,
     pub owner_pubkey: String,
-    pub max_bid_price: f64,
-    pub energy_requirements: f64,
+    pub max_bid_price: BigDecimal,
+    pub energy_requirements: BigDecimal,
     pub reputation_score: i32,
     pub status: String,
     pub last_seen: DateTime<Utc>,
@@ -36,9 +37,9 @@ pub struct Auction {
     pub id: i64,
     pub battery_id: i32,
     pub aggregator_id: Option<i32>,
-    pub energy_amount: f64,
-    pub reserve_price: f64,
-    pub final_price: Option<f64>,
+    pub energy_amount: BigDecimal,
+    pub reserve_price: BigDecimal,
+    pub final_price: Option<BigDecimal>,
     pub status: String,
     pub blockchain_tx_hash: Option<String>,
     pub started_at: DateTime<Utc>,
@@ -51,8 +52,8 @@ pub struct Bid {
     pub id: i64,
     pub auction_id: i64,
     pub aggregator_id: i32,
-    pub bid_price: f64,
-    pub energy_amount: f64,
+    pub bid_price: BigDecimal,
+    pub energy_amount: BigDecimal,
     pub status: String, // "pending", "accepted", "rejected"
     pub created_at: DateTime<Utc>,
     pub updated_at: DateTime<Utc>,
@@ -72,12 +73,12 @@ pub struct SystemMetrics {
 pub struct NewBattery {
     pub device_id: i32,
     pub owner_pubkey: String,
-    pub energy_total: f64,
-    pub percentage_for_sale: f64,
-    pub reserve_price: f64,
+    pub energy_total: BigDecimal,
+    pub percentage_for_sale: BigDecimal,
+    pub reserve_price: BigDecimal,
     pub health_status: i32,
-    pub voltage: f64,
-    pub discharge_rate: f64,
+    pub voltage: BigDecimal,
+    pub discharge_rate: BigDecimal,
     pub status: String,
 }
 
@@ -85,8 +86,8 @@ pub struct NewBattery {
 pub struct NewAggregator {
     pub device_id: i32,
     pub owner_pubkey: String,
-    pub max_bid_price: f64,
-    pub energy_requirements: f64,
+    pub max_bid_price: BigDecimal,
+    pub energy_requirements: BigDecimal,
     pub reputation_score: i32,
     pub status: String,
 }
@@ -94,8 +95,8 @@ pub struct NewAggregator {
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct NewAuction {
     pub battery_id: i32,
-    pub energy_amount: f64,
-    pub reserve_price: f64,
+    pub energy_amount: BigDecimal,
+    pub reserve_price: BigDecimal,
     pub status: String,
 }
 
@@ -103,7 +104,7 @@ pub struct NewAuction {
 pub struct NewBid {
     pub auction_id: i64,
     pub aggregator_id: i32,
-    pub bid_price: f64,
-    pub energy_amount: f64,
+    pub bid_price: BigDecimal,
+    pub energy_amount: BigDecimal,
     pub status: String,
 }
